Crude implem for `preferRelative` version
This version will transform paths in both direction (relative->absolute
and absolute->relative).

It works thus:
- it generates a normalized relative import path and a normalized
  absolute import
- if the generated import path matches the `preferRelative` regex passed
  in the options, use it, else use the absolute import.

It uses the file system to check whether the target of an absolute
import exists and there is part of the package, and could be replaced by
a relative import. This last bit could be improved upon.

It needs tests too.